refactor: ページコンポーネントを pages/ ディレクトリに移動
- ConfigPage, MetricsPage, SessionDetail を frontend/src/pages/ に移動
- SessionDetail を SessionPage にリネームして命名規則を統一
- App.tsx のインポートパスを更新
